moved shouldRender function into utilites folder and updated imports
