+++ ws-nextjs-app/pages/[service]/articles/handleArticleRoute.test.ts
@@ -1,6 +1,6 @@
 import pidginMediaArticleFixtureData from '#data/pidgin/articles/cvpde7nqj92o.json';
 import { GetServerSidePropsContext } from 'next';
-import * as shouldRender from './shouldRender';
+import * as shouldRender from '../../../utilities/shouldRender';
 import * as getPageDataModule from '../../../utilities/pageRequests/getPageData';
 import handleArticleRoute from './handleArticleRoute';
