+++ ws-nextjs-app/pages/[service]/articles/handleArticleRoute.ts
@@ -12,7 +12,7 @@ import { PageTypes } from '#app/models/types/global';
 import { ArticleMetadata } from '#app/models/types/optimo';
 import { getServerExperiments } from '#server/utilities/experimentHeader';
 import augmentWithDisclaimer from './augmentWithDisclaimer';
-import shouldRender from './shouldRender';
+import shouldRender from '../../../utilities/shouldRender';
 import getPageData from '../../../utilities/pageRequests/getPageData';
 
 // EXPERIMENT: Location based Topics Experiment
