+++ ws-nextjs-app/pages/[service]/homepage/handleHomepageRoute.ts
@@ -9,7 +9,7 @@ import getPathExtension from '#app/utilities/getPathExtension';
 import PageDataParams from '#app/models/types/pageDataParams';
 import handleError from '#app/routes/utils/handleError';
 import { getServerExperiments } from '#server/utilities/experimentHeader';
-import shouldRender from '../articles/shouldRender';
+import shouldRender from '../../../utilities/shouldRender';
 import getPageData from '../../../utilities/pageRequests/getPageData';
 
 const logger = nodeLogger(__filename);
